The nonzero percentage is now saved in the dataset files

+++ src/dataset_creation.cpp
@@ -51,7 +51,7 @@ std::vector<unsigned long long int> createDataset(unsigned long int tot_num, uns
 	{
 		// probabilistic choice between zero and non-zero
 		int rand_threshold = RAND_MAX * ((double) non_zero_percentage / 100);
-		for(int i=0; i < tot_num; i++)
+		for(unsigned long int i=0; i < tot_num; i++)
 		{	
 			if(rand() < rand_threshold)
 				// max_value/10 to really be sure that there won't be any overflows
